% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carselAndParrish.R
\name{cap_prior}
\alias{cap_prior}
\title{calculates the prior according to the Carsel and Parrish methodology}
\usage{
cap_prior(meas, theta)
}
\arguments{
\item{meas}{a vector of measurements}

\item{theta}{values for which to calculate the pdf}
}
\value{
the corresponding pdf
}
\description{
\code{cap_prior} calculates the prior according to the Carsel and Parrish methodology
}
