% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_porosity.R
\docType{data}
\name{df_porosity}
\alias{df_porosity}
\title{566 porosity values}
\format{A data frame with 566 rows and 2 variables:
\describe{
  \item{site_id}{id of the site, categorical}
  \item{val}{porosity value, non-dimensional fraction}
  ...
}}
\source{
\url{https://github.com/GeoStat-Bayesian/geostatDB}
}
\usage{
df_porosity
}
\description{
A dataset containing 566 porosity values from 6 different sites
}
\keyword{datasets}
