% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_wwhypda_K}
\alias{df_wwhypda_K}
\title{Hydraulic conductivity values from wwhypda}
\format{a data frame with 12505 rows and 39 variables:
\describe{
\item{val}{the numerical value for hydraulic conductivity}
\item{id_Measure}{the unique identifier for the measure}
\item{site_id}{unique identifier of the site where measurement were taken}
\item{rt_name}{name of rock type at the site}
\item{env_name}{name if environment type at the site}
}}
\source{
\url{http://wwhypda.unine.ch/}
}
\usage{
df_wwhypda_K
}
\description{
The dataset of hydraulic conductivity values reported in wwhypda
}
\keyword{datasets}

