% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwhypda2df_MySQL.R
\name{functionName}
\alias{functionName}
\title{query measurements from wwhypda database}
\usage{
functionName(par1)
}
\arguments{
\item{par1}{what is par1}

\item{par2}{what is par2}
}
\value{
a dataframe blabla
}
\description{
\code{functionName} does blabla
}
\examples{
write examples here
}

