% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_generalfromMeas_trans.R
\name{generalFromMeas_trans}
\alias{generalFromMeas_trans}
\title{general prior from measurements}
\usage{
generalFromMeas_trans(meas, eval_theta, site_specific_transform = F,
  type = c("Log_Normal", "Box-Cox", "Log Ratio", "Truncated_Normal"),
  arg.transform = c(-Inf, Inf), niter = 10^5, hierarchicalSigma = F,
  verbose = F)
}
\arguments{
\item{meas}{a dataframe containing measurements to assimilate, with fields val and site_id (see example)}

\item{eval_theta}{a vector of numerical values of informative prior evaluation points}

\item{niter}{an integer for the number of samples to use in the MCMC}

\item{hierarchicalSigma}{a boolean specifying whether the site-specific variance
is defined hierarchically by an inverse-gamma distribution (T) or by a prior (F)}

\item{verbose}{boolean indicating whether R should print information from the progress}
}
\value{
the pdf at values corresponding to theta
}
\description{
\code{generalFromMeas_trans} generates general priors from a set of measurements
from multiple sites
}
\examples{
theta_vect <- seq(from=-15,to=0,by=0.1)
df_meas <- data.frame(val=c(c(2,3,4),c(2,1),c(6,7,2,3)),
                     site_id=c(rep("a",3),rep("b",2),rep("c",4)))
generalFromMeas_trans(meas=df_meas,eval_theta=theta_vect)
}
