% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getData.R
\name{getData}
\alias{getData}
\title{extract data from wwhypda database}
\usage{
getData(rock_type = NULL, param = NULL, site = NULL)
}
\arguments{
\item{rock_type}{a character indicating rock type. if left blank, data from all rock types returned.}

\item{param}{a character indicating parameter. if left blank, data from all parameters returned.}

\item{site}{a character indicating site. if left blank, data from all sites returned.}
}
\value{
data queried from the wwhypda database as a dataframe
}
\description{
\code{getData} queries the wwhypda database according to specifications
of site name, rock type, and parameter of interest and returns data as a dataframe.
}
\examples{
my_data <- getData(rock_type = "Sandstone", param = "porosity")
head(my_data)

all_data <- getData()

lagenthal <- getData(site = "Langenthal")
dim(lagenthal)
}
