% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carselAndParrish.R
\name{goodness_of_fit}
\alias{goodness_of_fit}
\title{tests goodness of fit to normal distribution}
\usage{
goodness_of_fit(y)
}
\arguments{
\item{y}{a vector of samples for which to test the normality}
}
\value{
the Kolmogorov-Smirnov statistics
}
\description{
\code{goodness_of_fit} tests the closedness to the normal distribution
}

