% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carselAndParrish.R
\name{johnson_sb}
\alias{johnson_sb}
\title{log ratio transform}
\usage{
johnson_sb(x, a, b)
}
\arguments{
\item{x}{vector of original dataset to transform}

\item{a}{useless argument}

\item{b}{useless argument}
}
\value{
the transformed sample
}
\description{
\code{johnson_sb} is the log-ratio (2nd Johnson transform)
}
