% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carselAndParrish.R
\name{johnson_su}
\alias{johnson_su}
\title{hyperbolic arcsine transform}
\usage{
johnson_su(x, a, b)
}
\arguments{
\item{x}{vector of original dataset to transform}
}
\value{
the transformed sample
}
\description{
\code{johnson_su} is the hyperbolic arcsine - transform (3rd Johnson transform)
}
