% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_multiplot.R
\name{multiplot}
\alias{multiplot}
\title{multiplot function}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects}

\item{plotlist}{list of the ggplot objects}

\item{file}{filename}

\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored}
}
\value{
a plot
}
\description{
\code{multiplot} allows to combines multiple plots
}
