% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gPrior.R
\name{plot_gPrior}
\alias{plot_gPrior}
\title{plot general prior}
\usage{
plot_gPrior(res_gPrior, plotMeas = F)
}
\arguments{
\item{res_gPrior}{output from the generalFromMeas function}

\item{plotMeas}{boolean asking whether to plot the measurements}
}
\value{
a plot
}
\description{
\code{plot_gPrior} plot informative and non-informative priors
}

