% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hyperDist.R
\name{plot_hyperDist}
\alias{plot_hyperDist}
\title{plot prior and posterior distribution of hyperparameters}
\usage{
plot_hyperDist(res_gPrior)
}
\arguments{
\item{res_gPrior}{output from the generalFromMeas function}
}
\value{
a plot
}
\description{
\code{plot_hyperDist} plot prior and posterior distribution of hyperparameters
}

