% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gPrior.R
\name{plot_rPrior}
\alias{plot_rPrior}
\title{plot general prior}
\usage{
plot_rPrior(res_rPrior, plotMeas = F)
}
\arguments{
\item{res_rPrior}{output from the generalFromMeas function}

\item{plotMeas}{boolean asking whether to plot the measurements}
}
\value{
a plot
}
\description{
\code{plot_rPrior} plot informative and non-informative priors
}

