% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHyperDist.R
\name{plotHyperDist}
\alias{plotHyperDist}
\title{plot prior and posterior distribution of hyperparameters}
\usage{
plotHyperDist(resExPrior)
}
\arguments{
\item{resExPrior}{output from the genExPrior function}
}
\value{
a plot
}
\description{
\code{plotHyperDist} plot prior and posterior distribution of hyperparameters
}
\examples{
\donttest{
exdata <- data.frame(val=c(c(2,3,4),c(2,1),c(6,7,2,3)),
                  site_id=c(rep("a",3),rep("b",2),rep("c",4)),
                  x = c(c(2,3,4),c(2,3),c(2,2,3,3)),
                  y = c(c(2,2,3),c(3,2),c(2,3,2,3)))
ex_prior <- genExPrior(exdata=exdata,theta=seq(from=-5,to=5,by=1))
plotHyperDist(ex_prior)
}
}
