% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_regionalized.R
\name{regionalized}
\alias{regionalized}
\title{regionalized prior from multiple types of site-specific data}
\usage{
regionalized(data, eval_theta, niter = 10^5, hierarchicalSigma = F,
  verbose = F)
}
\arguments{
\item{data}{a dataframe containing data to assimilate,
with fields dat, site_id and type (see example)}

\item{eval_theta}{a vector of numerical values of informative prior evaluation points}

\item{niter}{an integer for the number of samples to use in the MCMC}

\item{hierarchicalSigma}{a boolean specifying whether the site-specific variance
is defined hierarchically by an inverse-gamma distribution (T) or by a prior (F)}

\item{verbose}{boolean indicating whether R should print information from the progress}
}
\value{
the pdf at values corresponding to theta
}
\description{
\code{regionalized} generates a regionalized prior
from site-specific data from multiple sites
and of multiple types
}
\examples{
theta_vect <- seq(from=-10,to=10,by=0.1)
data_site_A <- data.frame(dat=c(2,4),
                          site_id=rep('A',2),
                          type=c('bound.min','bound.max'))
data_site_B <- data.frame(dat=c(2,0.1),
                          site_id=rep('B',2),
                          type=c('moment.1','moment.2'))
data_site_C <- data.frame(dat=c(2,3,4),
                          site_id=rep('C',3),
                          type=c('meas','meas','meas'))
df_data <- rbind(data_site_A,data_site_B,data_site_C)
gPrior::regionalized(data=df_data,eval_theta=theta_vect)
}

