% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_viewInfo.R
\name{viewInfo}
\alias{viewInfo}
\title{view information on wwhypda database, such as list of rock types,
parameters, and sites.}
\usage{
viewInfo(password, db_name)
}
\arguments{
\item{password}{your password (optional)}

\item{db_name}{the name of the local wwhypda database (optional)}
}
\value{
list of information on wwhypda database
}
\description{
\code{viewInfo} returns information on wwhypda database
}
\examples{
info <- viewInfo()
}

