% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_wwhypdaConnect.R
\name{wwhypdaConnect}
\alias{wwhypdaConnect}
\title{extract data from wwhypda database}
\usage{
wwhypdaConnect(password, db_name)
}
\arguments{
\item{password}{your password}

\item{db_name}{the name of the local wwhypda database}
}
\value{
a connection to locally-stored wwhypda database
}
\description{
\code{wwhypdaConnect} connects to locally-stored wwhypda database with your password
}
\examples{
wwhypdaConnect(password='mypassword',db_name='mydbname')
}
