% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_generalfromMeas.R
\name{generalFromMeas}
\alias{generalFromMeas}
\title{general prior from measurements}
\usage{
generalFromMeas(meas, eval_theta, niter = 10^5, range_alpha = NULL,
  hierarchicalSigma = F, verbose = F)
}
\arguments{
\item{meas}{a dataframe containing measurements to assimilate, with fields val and site_id (see example)}

\item{eval_theta}{a vector of numerical values of informative prior evaluation points}

\item{niter}{(optional) an integer for the number of samples to use in the MCMC}

\item{range_alpha}{(optional) a vector of two values corresponding to
the lower and the upper bounds of the uniform distribution for alpha}

\item{hierarchicalSigma}{(optional) a boolean specifying whether the site-specific variance
is defined hierarchically by an inverse-gamma distribution (T) or by a prior (F)}

\item{verbose}{(optional) boolean indicating whether R should print information from the progress}
}
\value{
the pdf at values corresponding to theta
}
\description{
\code{generalFromMeas} generates general priors from a set of measurements
from multiple sites
}
\examples{
theta_vect <- seq(from=-10,to=10,by=0.1)
df_meas <- data.frame(val=c(c(2,3,4),c(2,1),c(6,7,2,3)),
                     site_id=c(rep("a",3),rep("b",2),rep("c",4)))
generalFromMeas(meas=df_meas,eval_theta=theta_vect)
}
