% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getData.R
\name{getData}
\alias{getData}
\title{extract data from wwhypda database}
\usage{
getData(password, db_name, rock_type = NULL, param = NULL, site = NULL,
  view_rocktypes = FALSE, view_params = FALSE, view_sites = FALSE)
}
\arguments{
\item{rockType}{a character indicating rock type. if left blank, data from all rock types returned.}

\item{param}{a character indicating parameter. if left blank, data from all parameters returned.}

\item{site}{a character indicating site. if left blank, data from all sites returned.}

\item{viewInfo}{logical; if TRUE, calls viewInfor(), returns list of rock types, parameters, and sites}
}
\value{
data queried from the wwhypda database as a dataframe. if viewInfo is selected, also returns output of viewInfo()
}
\description{
\code{getData} queries the wwhypda database according to specifications
of site name, rock type, and parameter of interest and returns data as a dataframe.
}
\examples{
my_data <- getData(rockType = "Sandstone", param = "porosity")
head(my_data)

all_data <- getData()

data_info <- getData(viewInfo=TRUE)

lagenthal <- getData(site = "Langenthal")
dim(lagenthal)

porosity_of_sandstone <- getData(rockType = "Sandstone, channel",
param = "porosity")

# this should raise an error:
site_unavailable <- getData(site = "U.C. Berkeley")
}
