% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carselAndParrish.R
\name{johnson_ln}
\alias{johnson_ln}
\title{log transform}
\usage{
johnson_ln(x, a, b)
}
\arguments{
\item{x}{vector of original dataset to transform}

\item{a}{useless argument}

\item{b}{useless argument}
}
\value{
the transformed sample
}
\description{
\code{johnson_ln} is the log-transform (1st Johnson transform)
}

