% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_multiplot.R
\name{multiplot}
\alias{multiplot}
\title{multiplot function}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\value{
a plot
}
\description{
\code{multiplot} allows to combines multiple plots
}
