% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExPrior.R
\name{plotExPrior}
\alias{plotExPrior}
\title{plot ex-situ prior}
\usage{
plotExPrior(resExPrior, plotExData = F)
}
\arguments{
\item{resExPrior}{output from the genExPrior function}

\item{plotExData}{boolean asking whether to additionaly plot the ex-situ data}
}
\value{
a plot
}
\description{
\code{plotExPrior} plot informative and non-informative priors
}
