% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHyperDist.R
\name{plotHyperDist}
\alias{plotHyperDist}
\title{plot prior and posterior distribution of hyperparameters}
\usage{
plotHyperDist(resExPrior)
}
\arguments{
\item{resExPrior}{output from the genExPrior function}
}
\value{
a plot
}
\description{
\code{plotHyperDist} plot prior and posterior distribution of hyperparameters
}
