% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_viewInfo.R
\name{viewInfo}
\alias{viewInfo}
\title{view information on wwhypda database: list of recognized rock types,
parameters, and sites.}
\usage{
viewInfo()
}
\arguments{
}
\value{
list of information on wwhypda database: rock types, parameters, sites
}
\description{
\code{viewInfo} returns information on wwhypda database
}
\examples{
info <- viewInfo()
}
