% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_wwhypdaConnect.R
\name{wwhypdaConnect}
\alias{wwhypdaConnect}
\title{extract data from wwhypda database}
\usage{
wwhypdaConnect(password)
}
\arguments{
\item{password}{your password}
}
\value{
a connection to locally-stored wwhypda database
}
\description{
\code{wwhypdaConnect} connects to locally-stored wwhypda database with your password
}
\examples{
wwhypdaConnect('mypassword')
}

